+++ Leomard/Services/PostService.swift
@@ -147,4 +147,11 @@ class PostService: Service {
             self.respond(result, completion)
         }
     }
+    
+    public func lock(post: Post, locked: Bool, completion: @escaping(Result<PostResponse, Error>) -> Void) {
+        let body = LockPost(locked: locked, postId: post.id)
+        requestHandler.makeApiRequest(host: SessionStorage.getInstance.getLemmyInstance(), request: "/post/lock", method: .post, body: body) { result in
+            self.respond(result, completion)
+        }
+    }
 }